# realestate/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real Estate API",
        "description": "API documentation for the Real Estate Laravel Application",
        "contact": {
            "email": "support@yourapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/property-types": {
            "get": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Get all property types",
                "operationId": "4732b47dc30b83e58224671df904f769",
                "responses": {
                    "200": {
                        "description": "List of property types"
                    },
                    "404": {
                        "description": "No property types found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Create a new property type",
                "operationId": "d246513d1c4b4073141e45bd05edec8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property type created"
                    },
                    "403": {
                        "description": "Only admin can create property types"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/property-types/{id}/properties": {
            "get": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Get all properties by property type ID",
                "operationId": "d7e5126b10e8a0072fe14c4a4a2a614b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of properties for the given property type"
                    },
                    "404": {
                        "description": "No properties found for this property type"
                    }
                }
            }
        },
        "/api/property-types/{id}": {
            "get": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Get a specific property type",
                "operationId": "171c3f177bf588137005565c8312f969",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property type found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Update a property type",
                "operationId": "4edd85b0d0fd8d16930403d5b8ad8a22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property type updated"
                    },
                    "403": {
                        "description": "Only admin can update property types"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Property Types"
                ],
                "summary": "Delete a property type",
                "operationId": "0d0f477e2ae67250924f8de19a0c54df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property type deleted"
                    },
                    "403": {
                        "description": "Only admin can delete property types"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}